Create README.md
Build a 3D globe app with the ArcGIS API for JavaScript

We are living in the sixth mass extinction. Half-Earth is a call to protect half the land and sea in order to manage sufficient habitat to slow down the species extinction crisis and ensure the long-term health of our planet. The Half-Earth Project is working to safeguard global biodiversity through detailed mapping of biodiversity, protected areas, and human activities.